HADOOP-2553 Don't make Long objects calculating hbase type hash codes
git-svn-id: https://svn.apache.org/repos/asf/lucene/hadoop/trunk@610698 13f79535-47bb-0310-9956-ffa450edef68

+++ src/contrib/hbase/src/java/org/apache/hadoop/hbase/HLog.java
@@ -332,13 +332,13 @@ public class HLog implements HConstants {
             // equal to the oldest pending region operation
             TreeSet<Long> sequenceNumbers =
               new TreeSet<Long>(this.outputfiles.headMap(
-                (oldestOutstandingSeqNum + Long.valueOf(1L))).keySet());
+                (Long.valueOf(oldestOutstandingSeqNum.longValue() + 1L))).keySet());
             // Now remove old log files (if any)
             if (LOG.isDebugEnabled()) {
               // Find region associated with oldest key -- helps debugging.
               Text oldestRegion = null;
               for (Map.Entry<Text, Long> e: this.lastSeqWritten.entrySet()) {
-                if (e.getValue().longValue() == oldestOutstandingSeqNum) {
+                if (e.getValue().longValue() == oldestOutstandingSeqNum.longValue()) {
                   oldestRegion = e.getKey();
                   break;
                 }
